refactor(site): split ejected skin generator
